requirements.txt
# Core dependencies
pandas==2.1.4
numpy==1.26.3
python-dotenv==1.0.0
pyyaml==6.0.1

# LLM and AI
openai==1.54.0
langchain
langchain-openai

# Data visualization and analysis
matplotlib==3.8.2
seaborn==0.13.1

# Utilities
python-dateutil==2.8.2
tqdm==4.66.1

# Testing
pytest==7.4.4
pytest-cov==4.1.0

# Optional observability
langfuse==2.20.0

# Development
black==24.1.1
flake8==7.0.0
mypy==1.8.0
